Add modular Streamlit app with page structure
Introduces a modular Streamlit application with separate pages for login, home, menu, insert, update, and viewer functionalities. The main entry point is now main.py, and the previous monolithic logic in streamlit_main.py is commented out in favor of the new structure. Requirements are updated to support the new modules.
