+++ src/pages/viewer_page.py
@@ -1,0 +1,4 @@
+import streamlit as st
+
+def viewer(con):
+    st.title("Visualizar Dados")
